requirements.txt
# ── Core RAG dependencies ─────────────────────────────────────────────────────
anthropic>=0.40.0
sentence-transformers>=2.7.0
chromadb>=0.5.0
pypdf>=4.0.0
pandas>=2.0.0
python-dotenv>=1.0.0

# ── Web UI ────────────────────────────────────────────────────────────────────
streamlit>=1.40.0

# ── Utilities ─────────────────────────────────────────────────────────────────
tqdm>=4.66.0
numpy>=1.26.0
